requirements.txt
tensorflow==2.10.0
PyYAML==6.0
sklearn==0.0
tqdm==4.64.1
matplotlib==3.3.4
numba==0.56.2
librosa==0.9.2